revert #106 due to failing revdepchecks
As it turns out, multiple R packages are building on
`get_logger_meta_variables` and their own wrappers, which make use
of the `list` format. Also running into scoping issues when not
loading the package, but using `logger::get_logger_meta_variables`.

Let's revisit after CRAN submission by creating a new version of this
function, update the internal layouts to use the new one, but let
others rely on the old/slow method. And we can also ping the other
package developers to switch over to the new method.