A lazier version of `get_logger_meta_variables()`
It's quite a bit faster even in the worst case at the top-level where parsing the call should be cheap:

```
bench::mark(get_logger_meta_env(), get_logger_meta_variables(), check = FALSE)[1:4]
# A tibble: 2 × 4
  expression                       min   median `itr/sec`
  <bch:expr>                  <bch:tm> <bch:tm>     <dbl>
1 get_logger_meta_env()         12.7µs   13.7µs    65636.
2 get_logger_meta_variables()   57.3µs   60.4µs    16034.
```

Fixes #105. @darcozig, if you think this is a reasonable approach, I'm happy to finish it off by adding some tests to ensure that I have actually implemented the delayed evaluation correctly.

IMO it's not necessary to export this function since it's primarily for internal usage, and it would be nice to retain the freedom to change its implementation if we discover ways to improve the peformance further in the future.

OTOH, I think it's probably worth moving to its own function, since its purpose is a little different to the rest of `layouts.R`