Use rlang for error wrapping, if available
This includes some general improvements to the erorr handling, e.g. using `conditionMessage()` instead of `$message` and `withCallingHandlers()` instead of `tryCatch()`.

I also made the hint conditional, since I've seen it a couple of times when it wasn't appropriate, and it's better not to send the user off chasing a red herring.